Added tasks, updated feedback and changed id for the appendix (#9)
**Pull Request Description**

Updated related exercises to be tasks, removed feedback from choose all
that apply question and updated the id for the appendix and reference to
it.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.