updated exercises and tasks ids and labels (#14)
**Pull Request Description**

Reformatted the chapter to reflect decisions regarding exercises, tasks
and associated id and label naming conventions.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.