Chapter4 fixing pretext (#26)
**Pull Request Description**

Cleaned up some pretext tags in the existing sections. Removed the
creating section from the chapter as the verifying section includes both
the linux and vscode versions.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.