added a polling question and added term tags (#34)
**Pull Request Description**

Updated the gitkit feedback section to by adding a poll for the amount
of time spent out of class and added term tags to enhance readability.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.