upstreaming chapter draft feedback using polls (#36)
**Pull Request Description**

Drafted the feedback section for the upstreaming chapter based upon the
staying synchronized chapter.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.