Adds content to FOSS Communities and Collaboration chapter (#20)
**Pull Request Description**

Fills in the content and exercises for the FOSS Communities and
Collaboration chapter.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.