updates workflow images to high res images (#40)
**Pull Request Description**

Updates the workflow diagram images to be high res images. 
Puts a copy of the source images powerpoint into the repository.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.