Content for Merge Conflicts Third Section (#44)
**Pull Request Description**

added text and exercises to merge conflicts third section

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.

Co-authored-by: Grant Braught <braught@dickinson.edu>