Text and Exercises for Last Three Sections of Merge Conflicts (#50)
…apter

**Pull Request Description**

added text and exercises to last three sections of merge conflicts
chapter

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.