restructure intro as instructor guide with sections for each chapter (#51)
**Pull Request Description**

The intro was restructured as an Instructor Guide because that's really
what it is. Separate sections were added for the instructor notes about
each chapter of the text.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.