Add draft of Chapter 2 instructor notes. (#53)
**Pull Request Description**

Places a draft of the chapter 2 instructor notes. The main thing yet to
be done is to link the Exercise Notes to the correct questions in the
chapter.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.