added links to qualtrics feedback surveys (#57)
**Pull Request Description**

Changed the feedback from polls within the chapters to links to external
surveys. Left the original polling questions in as comments.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.