Adds instructor notes for chapters 3, 4, 5 (#60)
**Pull Request Description**

Adds in the instructor notes for the last 3 chapters. These still need
to have their questions linked to the proper exercises in the text.

Also updates the frontmatter to have an overview instead of a preface.
The overview text was revised. Acknowledgements were added to the
frontmatter as well.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.