adds nsf acknowledgement and pre/post survey links (#63)
**Pull Request Description**

Adds the NSF acknowledgement to the frontmatter.

Adds sections at the very beginning and very end that describe and link
to the pre/post surveys.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.