makes exercise/task refs into links in instructor guide (#64)
**Pull Request Description**

Changes the use of `title` tags to `p` tags within the `li` tags that
make up the comments on different exercises. When the `title` tag was
used the exercise/task name was displayed but it was not a link. Using a
`p` tag allows them to be links.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.