try changing the document-id to gitkit linux (#65)
**Pull Request Description**

It would be nice for the document id to reflect the version, so this PR
is an attempt to change the `document-id` in the `dockinfo.ptx` file for
the linux version of the text. The question is whether Runestone will
pick it up when the repo is changed or if we'll need to create a new
book using the new id?

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.