update questions to fix issues with kit commits (#66)
**Pull Request Description**

Fixes some questions in chapter 3 that depend upon the dates and order
of commits. Deploying the kit adds a new commit, so the date, author,
etc of the most recent commit changes. The questions here have been
revised to account for this.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.