creates a runestone target for build (#67)
**Pull Request Description**

Creates a single `runestone` target that builds the linux desktop
version of the book. The fork for the VSCode version of the text will
then need to change this target to point to the correct publication
file.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.