adds instructor materials related to Runestone (#68)
**Pull Request Description**

Add material to the instructor guide about the different editions (Linux
Desktop, VSCode) with links to the different editions. Also adds
information and links to information about creating courses and
assignments in Runestone.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.