updates the readme to reflect published versions (#70)
**Pull Request Description**

- Notes that this is the repo for the source.
- Adds pointers to the GitKit editions on Runestone.
- Adds info about publishing / updating versions.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.