Survey update (#78)
**Pull Request Description**

Updates the pre-survey section and the instructor guide for the new
survey that allows students to consent to providing their responses to
all GitKit exercises.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.