move runestone output to unique dir for linux edition (#80)
**Pull Request Description**

Changes the output directory for the runestone linux edition to a
directory that is unique to that edition. This is an attempt to fix a
build issue on Runestone where the VSCode edition was still building the
Linux edition.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.