Updates upstreaming (#119)
**Pull Request Description**

_Describe what your Pull Request does here_
This PR contains updates to Ch. 3 - Upstreaming Changes.

_Include a "Closes #" line for any issues that your PR closes._

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.