escape . in regular expressions (#136)
**Pull Request Description**

Escapes the . in regular expression checks for `user.name` and
`user.email`.

Closes #121 
Closes #122 

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.