Chapter 2 Revisions (#146)
**Pull Request Description**

Completes all revisions for Chapter 2 - except for 2.8.1.c - consider
blurring the username and repo name in the URL in the image.

Closes #90

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.